requirements-enhanced.txt
# Requirements for Face Recognition System

## Core Dependencies (Already in requirements.txt)
boto3>=1.34.0
botocore>=1.34.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.5.0
pydantic-settings>=2.1.0

## New Dependencies for Enhanced Features

# Redis Caching (for sub-50ms latency)
redis>=5.0.0
hiredis>=2.2.3  # C parser for better performance

# Image Quality Validation (anti-spoofing)
opencv-python-headless>=4.8.1.78
numpy>=1.24.0
Pillow>=10.1.0

# Additional utilities
python-dotenv>=1.0.0
